fix: use promices on main function
