+++ app/lib/misc/helpers.ts
@@ -1,0 +1,10 @@
+export function toResult(
+  element: any,
+  innerSelector: string,
+  extractType: "textContent" | "href"
+) {
+  const e = element.querySelector(`${innerSelector}`);
+  if (e) {
+    return e[extractType];
+  }
+}
